remove dummy file and spec
