--- src/js/components/dummy-for-test.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-export default React.createClass({
-
-  displayName: 'DummyForTest',
-
-  render() {
-    return(
-      <div>
-        <span>Dummy For Test</span>
-        <button className='my-button' onClick={this.props.onClick}/>
-      </div>
-    );
-  }
-
-});
